Changed Directory Structure
This makes it a bit easier to scan through files.
