+++ src/components/Collage/style.css
