+++ src/components/Footer/NowPlaying.tsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom';
 import Artwork from '../Artwork';
 import { join, unique } from '../../utils';
 import { album, artist } from '../../paths';
-import { Song } from '../../containers/Footer/query';
+import { Song } from '../../containers/Footer/enhancers/query';
 import * as styles from './NowPlaying.css';
 import DefaultCover from '../icons/DefaultCover';
