+++ src/components/PlaybackArtwork/style.css
