+++ src/components/Sidebar/index.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
-import * as styles from './Sidebar.css';
+import * as styles from './style.css';
 import { NavLink } from 'react-router-dom';
 
-import { albums, artists, home, queue, search, songs } from '../paths';
+import { albums, artists, home, queue, search, songs } from '../../paths';
 
 interface Props {
   className: string;
