+++ src/components/Sidebar/style.css
@@ -1,4 +1,4 @@
-@import '../colors.css';
+@import '../../colors.css';
 
 .link {
   /*
