+++ src/components/SongCollage/index.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 
-import Artwork from './Artwork';
-import DefaultCover from './icons/DefaultCover';
-import Collage from './Collage';
+import Artwork from '../Artwork';
+import DefaultCover from '../icons/DefaultCover';
+import Collage from '../Collage';
 
 export function must<T>(input: T | undefined): T {
   if (!input) {
