+++ src/components/VolumeSlider/styles.css
@@ -1,5 +1,5 @@
-@import '../colors.css';
-@import '../sizes.css';
+@import '../../colors.css';
+@import '../../sizes.css';
 
 :root {
   --thumb-size: 12px;
