+++ src/containers/Footer/index.ts
@@ -1,9 +1,12 @@
 import { ComponentType } from 'react';
 
-import Footer from './Footer';
-import { reduxEnhancer } from './redux';
-import { graphqlEnhancer } from './query';
-import { likeMutationEnhancer, playSongMutationEnhancer } from './mutate';
+import Footer from './component';
+import { reduxEnhancer } from './enhancers/redux';
+import { graphqlEnhancer } from './enhancers/query';
+import {
+  likeMutationEnhancer,
+  playSongMutationEnhancer,
+} from './enhancers/mutate';
 
 export interface InputProps {
   // A class applied to the component's container container element.
