+++ src/containers/Queue/component/styles.css
@@ -1,4 +1,4 @@
-@import '../../sizes.css';
+@import '../../../sizes.css';
 
 .heading {
   color: #ffffff;
